+++ AppForTabs/Views/FretView/FretView.swift
@@ -8,8 +8,6 @@
 import SwiftUI
 
 struct FretView: View {
-    @State var pageIndex = 0
-    private let page: [Int] = [0,1,2,3]
     private let dotAppearance = UIPageControl.appearance()
     
     var body: some View {
